requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
de-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.0.0/de_core_news_lg-3.0.0-py3-none-any.whl
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.0.0/de_core_news_md-3.0.0-py3-none-any.whl
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0/de_core_news_sm-3.0.0-py3-none-any.whl
de-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/de_dep_news_trf-3.0.0/de_dep_news_trf-3.0.0-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.0.12
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pathy==0.5.2
preshed==3.0.5
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-alignments==0.8.3
spacy-legacy==3.0.5
spacy-transformers==1.0.2
srsly==2.4.1
thinc==8.0.3
tokenizers==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
xlrd==2.0.1
